refactor(VExpansionPanelItem): deprecate panel-content, clean-up styles
