+++ packages/vuetify/src/components/VExpansionPanel/VExpansionPanelItem.ts
@@ -1,0 +1,140 @@
+// Components
+import VExpansionPanel from './VExpansionPanel'
+import { VExpandTransition } from '../transitions'
+import VIcon from '../VIcon'
+
+// Mixins
+import { factory as GroupableFactory } from '../../mixins/groupable'
+import Bootable from '../../mixins/bootable'
+import Rippleable from '../../mixins/rippleable'
+
+// Utilities
+import mixins, { ExtractVue } from '../../util/mixins'
+import { consoleWarn } from '../../util/console'
+
+// Types
+import { VNode } from 'vue'
+
+const baseMixins = mixins(
+  Bootable,
+  GroupableFactory('expansionPanel'),
+  Rippleable
+)
+
+interface options extends ExtractVue<typeof baseMixins> {
+  $el: HTMLElement
+  expansionPanel: InstanceType<typeof VExpansionPanel>
+}
+
+/* @vue/component */
+export default baseMixins.extend<options>().extend({
+  name: 'v-expansion-panel-item',
+
+  props: {
+    disabled: Boolean,
+    readonly: Boolean,
+    expandIcon: {
+      type: String,
+      default: '$vuetify.icons.expand'
+    },
+    hideActions: Boolean,
+    ripple: {
+      type: [Boolean, Object],
+      default: false
+    }
+  },
+
+  data: () => ({
+    height: 'auto'
+  }),
+
+  computed: {
+    classes (): object {
+      return {
+        'v-expansion-panel-item--active': this.isActive,
+        'v-expansion-panel-item--disabled': this.isDisabled,
+        ...this.groupClasses
+      }
+    },
+    isDisabled (): boolean {
+      return this.expansionPanel.disabled || this.disabled
+    },
+    isReadonly (): boolean {
+      return this.expansionPanel.readonly || this.readonly
+    }
+  },
+
+  methods: {
+    onKeydown (e: KeyboardEvent) {
+      // Ensure element is the activeElement
+      if (
+        e.keyCode === 13 &&
+        this.$el === document.activeElement
+      ) this.click(e)
+    },
+    click (e: MouseEvent | KeyboardEvent) {
+      if (e.detail) this.$el.blur()
+
+      this.$emit('click', e)
+
+      this.isReadonly || this.toggle()
+    },
+    genBody () {
+      return this.$createElement('div', {
+        ref: 'body',
+        class: 'v-expansion-panel-item__body',
+        directives: [{
+          name: 'show',
+          value: this.isActive
+        }]
+      }, this.showLazyContent(this.$slots.default))
+    },
+    genHeader () {
+      const children = [...(this.$slots.header || [])]
+
+      if (!this.hideActions) children.push(this.genIcon())
+
+      return this.$createElement('div', {
+        staticClass: 'v-expansion-panel-item__header',
+        directives: [{
+          name: 'ripple',
+          value: this.ripple
+        }],
+        on: { click: this.click }
+      }, children)
+    },
+    genIcon () {
+      const icon = this.$slots.actions ||
+        [this.$createElement(VIcon, this.expandIcon)]
+
+      return this.$createElement('transition', {
+        attrs: { name: 'fade-transition' }
+      }, [
+        this.$createElement('div', {
+          staticClass: 'v-expansion-panel-item__header__icon',
+          directives: [{
+            name: 'show',
+            value: !this.isDisabled
+          }]
+        }, icon)
+      ])
+    }
+  },
+
+  render (h): VNode {
+    return h('li', {
+      staticClass: 'v-expansion-panel-item',
+      class: this.classes,
+      attrs: {
+        tabindex: this.isReadonly || this.isDisabled ? null : 0,
+        'aria-expanded': String(!!this.isActive)
+      },
+      on: {
+        keydown: this.onKeydown
+      }
+    }, [
+      this.$slots.header && this.genHeader(),
+      h(VExpandTransition, [this.genBody()])
+    ])
+  }
+})
