+++ packages/vuetify/src/components/VExpansionPanel/_mixins.sass
@@ -1,13 +1,13 @@
 @mixin expansion-panel-theme($material)
-  .v-expansion-panel__container
+  .v-expansion-panel-item
     background-color: map-get($material, 'cards')
     color: map-deep-get($material, 'text', 'primary')
 
     &:after
       background-color: map-get($material, 'dividers')
 
-    .v-expansion-panel__header
-      .v-expansion-panel__header__icon
+    .v-expansion-panel-item__header
+      .v-expansion-panel-item__header__icon
         .v-icon
           color: map-deep-get($material, 'icons', 'active')
